refactor(dashboard): split lighthouse tooling
